requirements.txt
# API Clients
openai==1.0.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
apscheduler==3.10.4
pandas==2.0.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Other
async_timeout==4.0.2
apscheduler==3.10.4
